- Вынес некоторые функции в хелперы - Вынес большие темплейты в отдельную папку - Убрал неиспользуемые вызовы рендера фильмов, которые идут в доп. задании
