+++ src/constants.js
@@ -7,4 +7,6 @@ const SortType = {
   BY_RATING: 'by_rating',
 };
 
-export {POPUP_BUTTON_ACTIVE_CLASS_NAME, CARD_BUTTON_ACTIVE_CLASS_NAME, SortType};
+const MAX_SHORT_DESCRIPTION_LENGTH = 140;
+
+export {POPUP_BUTTON_ACTIVE_CLASS_NAME, CARD_BUTTON_ACTIVE_CLASS_NAME, SortType, MAX_SHORT_DESCRIPTION_LENGTH};
