+++ src/view/sort-view.js
@@ -1,12 +1,6 @@
 import AbstractView from './abstract-view';
-import {SortType} from '../constants';
 import {removeClassFromElementList} from '../utils/common';
-
-const createSortTemplate = () => `<ul class="sort">
-    <li><a href="#" class="sort__button sort__button--active" data-sort-type="${SortType.DEFAULT}">Sort by default</a></li>
-    <li><a href="#" class="sort__button" data-sort-type="${SortType.BY_DATE}">Sort by date</a></li>
-    <li><a href="#" class="sort__button" data-sort-type="${SortType.BY_RATING}">Sort by rating</a></li>
-  </ul>`;
+import {createSortTemplate} from './templates/sort-template';
 
 export default class SortView extends AbstractView {
   get template() {
