+++ src/view/templates/sort-template.js
@@ -1,0 +1,9 @@
+import {SortType} from '../../constants';
+
+const createSortTemplate = () => `<ul class="sort">
+    <li><a href="#" class="sort__button sort__button--active" data-sort-type="${SortType.DEFAULT}">Sort by default</a></li>
+    <li><a href="#" class="sort__button" data-sort-type="${SortType.BY_DATE}">Sort by date</a></li>
+    <li><a href="#" class="sort__button" data-sort-type="${SortType.BY_RATING}">Sort by rating</a></li>
+  </ul>`;
+
+export {createSortTemplate};
